+++ config/test/test_config.global.php
@@ -8,8 +8,10 @@ use GuzzleHttp\Client;
 use Laminas\ConfigAggregator\ConfigAggregator;
 use Laminas\Diactoros\Response\EmptyResponse;
 use Laminas\ServiceManager\Factory\InvokableFactory;
+use League\Event\EventDispatcher;
 use Monolog\Level;
 use PHPUnit\Runner\Version;
+use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
@@ -20,7 +22,12 @@ use SebastianBergmann\CodeCoverage\Report\Html\Facade as Html;
 use SebastianBergmann\CodeCoverage\Report\PHP;
 use SebastianBergmann\CodeCoverage\Report\Xml\Facade as Xml;
 use Shlinkio\Shlink\Common\Logger\LoggerType;
+use Symfony\Component\Console\Application;
+use Symfony\Component\Console\Event\ConsoleCommandEvent;
+use Symfony\Component\Console\Event\ConsoleTerminateEvent;
+use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
+use function file_exists;
 use function Laminas\Stratigility\middleware;
 use function Shlinkio\Shlink\Config\env;
 use function sprintf;
@@ -30,13 +37,39 @@ use const ShlinkioTest\Shlink\SWOOLE_TESTING_HOST;
 use const ShlinkioTest\Shlink\SWOOLE_TESTING_PORT;
 
 $isApiTest = env('TEST_ENV') === 'api';
+$isCliTest = env('TEST_ENV') === 'cli';
+$isE2eTest = $isApiTest || $isCliTest;
 $generateCoverage = env('GENERATE_COVERAGE') === 'yes';
-if ($isApiTest && $generateCoverage) {
+
+$coverage = null;
+if ($isE2eTest && $generateCoverage) {
     $filter = new Filter();
     $filter->includeDirectory(__DIR__ . '/../../module/Core/src');
-    $filter->includeDirectory(__DIR__ . '/../../module/Rest/src');
+    $filter->includeDirectory(__DIR__ . '/../../module/' . ($isApiTest ? 'Rest' : 'CLI') . '/src');
     $coverage = new CodeCoverage((new Selector())->forLineCoverage($filter), $filter);
 }
+
+/**
+ * @param 'api'|'cli' $type
+ */
+$exportCoverage = static function (string $type = 'api') use (&$coverage): void {
+    if ($coverage === null) {
+        return;
+    }
+
+    $basePath = __DIR__ . '/../../build/coverage-' . $type;
+    $covPath = $basePath . '.cov';
+
+    // Every CLI test runs on its own process and dumps the coverage afterwards.
+    // Try to load it and merge it, so that we end up with the whole coverage at the end.
+    if ($type === 'cli' && file_exists($covPath)) {
+        $coverage->merge(require $covPath);
+    }
+
+    (new PHP())->process($coverage, $covPath);
+    (new Xml(Version::getVersionString()))->process($coverage, $basePath . '/coverage-xml');
+    (new Html())->process($coverage, $basePath . '/coverage-html');
+};
 
 $buildDbConnection = static function (): array {
     $driver = env('DB_DRIVER', 'sqlite');
@@ -113,17 +146,10 @@ return [
         [
             'name' => 'dump_coverage',
             'path' => '/api-tests/stop-coverage',
-            'middleware' => middleware(static function () use (&$coverage) {
+            'middleware' => middleware(static function () use ($exportCoverage) {
                 // TODO I have tried moving this block to a listener so that it's invoked automatically,
                 //      but then the coverage is generated empty ¯\_(ツ)_/¯
-                if ($coverage) { // @phpstan-ignore-line
-                    $basePath = __DIR__ . '/../../build/coverage-api';
-
-                    (new PHP())->process($coverage, $basePath . '.cov');
-                    (new Xml(Version::getVersionString()))->process($coverage, $basePath . '/coverage-xml');
-                    (new Html())->process($coverage, $basePath . '/coverage-html');
-                }
-
+                $exportCoverage();
                 return new EmptyResponse();
             }),
             'allowed_methods' => ['GET'],
@@ -164,6 +190,62 @@ return [
         'factories' => [
             TestUtils\Helper\TestHelper::class => InvokableFactory::class,
         ],
+        'delegators' => $isCliTest ? [
+            Application::class => [
+                static function (
+                    ContainerInterface $c,
+                    string $serviceName,
+                    callable $callback,
+                ) use (
+                    &$coverage,
+                    $exportCoverage,
+                ) {
+                    /** @var Application $app */
+                    $app = $callback();
+                    $wrappedEventDispatcher = new EventDispatcher();
+
+                    // When the command starts, start collecting coverage
+                    $wrappedEventDispatcher->subscribeTo(
+                        ConsoleCommandEvent::class,
+                        static function () use (&$coverage): void {
+                            $id = env('COVERAGE_ID');
+                            if ($id === null) {
+                                return;
+                            }
+
+                            $coverage?->start($id);
+                        },
+                    );
+                    // When the command ends, stop collecting coverage
+                    $wrappedEventDispatcher->subscribeTo(
+                        ConsoleTerminateEvent::class,
+                        static function () use (&$coverage, $exportCoverage): void {
+                            $id = env('COVERAGE_ID');
+                            if ($id === null) {
+                                return;
+                            }
+
+                            $coverage?->stop();
+                            $exportCoverage('cli');
+                        },
+                    );
+
+                    $app->setDispatcher(new class ($wrappedEventDispatcher) implements EventDispatcherInterface {
+                        public function __construct(private EventDispatcher $wrappedDispatcher)
+                        {
+                        }
+
+                        public function dispatch(object $event, ?string $eventName = null): object
+                        {
+                            $this->wrappedDispatcher->dispatch($event);
+                            return $event;
+                        }
+                    });
+
+                    return $app;
+                },
+            ],
+        ] : [],
     ],
 
     'entity_manager' => [
@@ -172,6 +254,7 @@ return [
 
     'data_fixtures' => [
         'paths' => [
+            // TODO These are used for CLI tests too, so maybe should be somewhere else
             __DIR__ . '/../../module/Rest/test-api/Fixtures',
         ],
     ],